requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==3.2.0
python-taiga==1.0.0
